requirements.txt

pyodbc
django-mssql-backend


asgiref==3.7.2
aws-psycopg2
azure-core
azure-functions
azure-identity
azure-storage-blob
certifi
cffi
charset-normalizer
configparser
cryptography
defusedxml
distlib
django-cors-headers
django-environ
django-storages
djangorestframework
dnspython
filelock
httpie
idna
isodate
markdown-it-py
mdurl
msal
msal-extensions
multidict
pipenv
platformdirs
portalocker

pycparser
pygments

pysocks
pytz
requests
requests-toolbelt
rich
six
sqlparse
typing_extensions
urllib3
virtualenv
virtualenv-clone
whitenoise

django-storages[azure]
django==4.1
django-environ
psycopg2-binary==2.9.1

djangorestframework
pillow 

django-extensions
drf-yasg
djangorestframework-simplejwt
pyjwt

--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.0.0+cpu
--find-links https://download.pytorch.org/whl/torch_stable.html
torchvision==0.15.0+cpu
--find-links https://download.pytorch.org/whl/torch_stable.html
torchaudio==2.0.0+cpu
--find-links https://download.pytorch.org/whl/torch_stable.html
numpy==1.25.2